compiles

+++ shared.h
@@ -4,7 +4,9 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-extern int wanted_position;
+#include "SevSeg.h"
+
+extern float wanted_position;
 extern int current_position_mix_valve;
 extern int temp_correction;
 extern SevSeg sevseg; //Instantiate a seven segment object
